+++ bin/cache-key.ts
@@ -1,1 +1,1 @@
-export default '59151a7a3c46dff3'
+export default '6bec8ef58c162ee4'